npc systems init
